added AvailablePackageSummaryFromChart() to pkg so it can be reused by multiple plugins
