+++ cmd/kubeapps-apis/plugins/fluxv2/packages/v1alpha1/repo_test.go
@@ -1531,6 +1531,7 @@ var valid_index_charts_spec = []testSpecChartWithFile{
 
 var valid_index_package_summaries = []*corev1.AvailablePackageSummary{
 	{
+		Name:        "acs-engine-autoscaler",
 		DisplayName: "acs-engine-autoscaler",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "2.1.1",
@@ -1543,8 +1544,10 @@ var valid_index_package_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "default", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	},
 	{
+		Name:        "wordpress",
 		DisplayName: "wordpress",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "0.7.5",
@@ -1557,10 +1560,12 @@ var valid_index_package_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "default", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	},
 }
 
 var cert_manager_summary = &corev1.AvailablePackageSummary{
+	Name:        "cert-manager",
 	DisplayName: "cert-manager",
 	LatestVersion: &corev1.PackageAppVersion{
 		PkgVersion: "v1.4.0",
@@ -1573,9 +1578,11 @@ var cert_manager_summary = &corev1.AvailablePackageSummary{
 		Context:    &corev1.Context{Namespace: "ns1", Cluster: KubeappsCluster},
 		Plugin:     fluxPlugin,
 	},
+	Categories: []string{""},
 }
 
 var elasticsearch_summary = &corev1.AvailablePackageSummary{
+	Name:        "elasticsearch",
 	DisplayName: "elasticsearch",
 	LatestVersion: &corev1.PackageAppVersion{
 		PkgVersion: "15.5.0",
@@ -1588,9 +1595,11 @@ var elasticsearch_summary = &corev1.AvailablePackageSummary{
 		Context:    &corev1.Context{Namespace: "default", Cluster: KubeappsCluster},
 		Plugin:     fluxPlugin,
 	},
+	Categories: []string{"Analytics"},
 }
 
 var ghost_summary = &corev1.AvailablePackageSummary{
+	Name:        "ghost",
 	DisplayName: "ghost",
 	LatestVersion: &corev1.PackageAppVersion{
 		PkgVersion: "13.0.14",
@@ -1603,6 +1612,7 @@ var ghost_summary = &corev1.AvailablePackageSummary{
 		Context:    &corev1.Context{Namespace: "default", Cluster: KubeappsCluster},
 		Plugin:     fluxPlugin,
 	},
+	Categories: []string{"CMS"},
 }
 
 var index_with_categories_summaries = []*corev1.AvailablePackageSummary{
@@ -1612,6 +1622,7 @@ var index_with_categories_summaries = []*corev1.AvailablePackageSummary{
 
 var index_before_update_summaries = []*corev1.AvailablePackageSummary{
 	{
+		Name:        "alpine",
 		DisplayName: "alpine",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "0.2.0",
@@ -1623,8 +1634,10 @@ var index_before_update_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "ns2", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	},
 	{
+		Name:        "nginx",
 		DisplayName: "nginx",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "1.1.0",
@@ -1636,11 +1649,13 @@ var index_before_update_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "ns2", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	},
 }
 
 var index_after_update_summaries = []*corev1.AvailablePackageSummary{
 	{
+		Name:        "alpine",
 		DisplayName: "alpine",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "0.3.0",
@@ -1652,8 +1667,10 @@ var index_after_update_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "ns2", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	},
 	{
+		Name:        "nginx",
 		DisplayName: "nginx",
 		LatestVersion: &corev1.PackageAppVersion{
 			PkgVersion: "1.1.0",
@@ -1665,4 +1682,5 @@ var index_after_update_summaries = []*corev1.AvailablePackageSummary{
 			Context:    &corev1.Context{Namespace: "ns2", Cluster: KubeappsCluster},
 			Plugin:     fluxPlugin,
 		},
+		Categories: []string{""},
 	}}
