+++ cmd/kubeapps-apis/plugins/pkg/packageutils/packageutils.go
@@ -15,6 +15,7 @@ package packageutils
 import (
 	"github.com/Masterminds/semver"
 	corev1 "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/core/packages/v1alpha1"
+	plugins "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/core/plugins/v1alpha1"
 	"github.com/kubeapps/kubeapps/pkg/chart/models"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -123,3 +124,53 @@ func IsValidChart(chart *models.Chart) (bool, error) {
 	}
 	return true, nil
 }
+
+// AvailablePackageSummaryFromChart builds an AvailablePackageSummary from a Chart
+func AvailablePackageSummaryFromChart(chart *models.Chart, plugin *plugins.Plugin) (*corev1.AvailablePackageSummary, error) {
+	pkg := &corev1.AvailablePackageSummary{}
+
+	isValid, err := IsValidChart(chart)
+	if !isValid || err != nil {
+		return nil, status.Errorf(codes.Internal, "invalid chart: %s", err.Error())
+	}
+
+	pkg.Name = chart.Name
+	// Helm's Chart.yaml (and hence our model) does not include a separate
+	// display name, so the chart name is also used here.
+	pkg.DisplayName = chart.Name
+	pkg.IconUrl = chart.Icon
+	pkg.ShortDescription = chart.Description
+	// TODO (gfichtenholt) I think when chart.Category is "" (i.e. missing)
+	// we should return nil rather than []string{""}. For now keeping the logic
+	// as is, so as not to break existing unit tests
+	pkg.Categories = []string{chart.Category}
+
+	pkg.AvailablePackageRef = &corev1.AvailablePackageReference{
+		Identifier: chart.ID,
+		Plugin:     plugin,
+	}
+	pkg.AvailablePackageRef.Context = &corev1.Context{Namespace: chart.Repo.Namespace}
+
+	if chart.ChartVersions != nil || len(chart.ChartVersions) != 0 {
+		pkg.LatestVersion = &corev1.PackageAppVersion{
+			PkgVersion: chart.ChartVersions[0].Version,
+			AppVersion: chart.ChartVersions[0].AppVersion,
+		}
+	}
+
+	return pkg, nil
+}
+
+// @gfichtenholt: I really wanted to put helm plugin's implementation of AvailablePackageDetailFromChart()
+// here, and use it in flux plugin as well. But I found out a couple of flaws in the implementation and decided
+// against it. Namely:
+// 1. This assumption:
+//    // We assume that chart.ChartVersions[0] will always contain either: the latest version or
+//    // the specified version
+//  This is a hack and forces the caller to prepate the input argument in a certain way such that
+//  chart.ChartVersions[0] will be the desired version. I other words the abstraction between the caller and
+//  the call site is completely broken here. Yuck.
+// 2. IMHO, it would have been better to get most of the detail info, including the current version out of
+//   parsed chart YAML file, which this implementation chooses to ignore.
+// I did consider using flux's implementation of AvailablePackageDetailFromChart but did not feel comfortable
+// chaning helm plugin to use it before talking to @minelson
